# swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: TelPro - FileMaker App (Hive)
# during dev, should point to your local machine
#host: 52.34.207.5:5058
host: localhost:5000
...
basePath: /
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /api/salesrep:
    # binds a127 app logic to a route
    x-swagger-router-controller: common_ctrl
    post:
      tags: ['Common Module']
      description: get sales rep by company id
      summary: get sales rep by company id
      # used as the method name of the controller
      operationId: SalesRep
      parameters:
        - name: body
          in: body
          description: Get sales rep
          schema:
           $ref: '#/definitions/SalesRep'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /api/getInvoice:
    # binds a127 app logic to a route
    x-swagger-router-controller: invoices_ctrl
    post:
      tags: ['Invoice Module']
      description: get list of Invoice
      summary: get list of Invoice
      # used as the method name of the controller
      operationId: getContactInvoice
      parameters:
        - name: body
          in: body
          description: Get list of Invoices
          schema:
           $ref: '#/definitions/commonparams'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/addInvoice:
    # binds a127 app logic to a route
    x-swagger-router-controller: invoices_ctrl
    post:
      tags: ['Invoice Module']
      description: add invoice
      summary: add Invoice
      # used as the method name of the controller
      operationId: addInvoice
      parameters:
        - name: body
          in: body
          description: add Invoice
          schema:
           $ref: '#/definitions/addInvoice'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/updateInvoice:
    # binds a127 app logic to a route
    x-swagger-router-controller: invoices_ctrl
    post:
      tags: ['Invoice Module']
      description: update invoice
      summary: update invoice
      # used as the method name of the controller
      operationId: updateInvoice
      parameters:
        - name: body
          in: body
          description: update invoice
          schema:
           $ref: '#/definitions/updateInvoice'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /api/getInvoiceDetails:
    # binds a127 app logic to a route
    x-swagger-router-controller: invoices_ctrl
    post:
      tags: ['Invoice Module']
      description: get  Invoice details
      summary: get  Invoice details
      # used as the method name of the controller
      operationId: getInvoiceDetails
      parameters:
        - name: body
          in: body
          description: get  Invoice details
          schema:
           $ref: '#/definitions/getInvoiceDetails'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"              
  /api/deleteInvoice:
    # binds a127 app logic to a route
    x-swagger-router-controller: invoices_ctrl
    post:
      tags: ['Invoice Module']
      description: delete Invoice
      summary: delete Invoice
      # used as the method name of the controller
      operationId: deleteInvoice
      parameters:
        - name: body
          in: body
          description: delete Invoice
          schema:
           $ref: '#/definitions/deleteInvoice'
          required: true        
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/appResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"                                      
...
definitions:
...
  appResponse:
    required:
      - code
    properties:
      code:
        type: number
      message:
        type: string
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
...
  SalesRep:
    type: object
    properties: 
      companyId:
        type: string
      firstname:
        type: string
...
  commonparams:
    type: object
    properties:
      companyId:
        type: string
      contactId:
        type: string
...
  addInvoice:
    type: object
    properties:  
      contactId:
        type: string
      projectId:
        type: string
      individualId:
        type: string
      opportunityId:
        type: string
      orderId:
        type: string
      estimateId:
        type: string
      poNumber:
        type: string
      title:
        type: string
      createdDate:
        type: string
      dueDate:
        type: string
      invoiceNumber:
        type: string
      invoiceItem:
        type: string
      terms:
        type: number
      class:
        type: number
      salesRep:
        type: string
      customerMessage:
        type: number
      memo:
        type: string
      total:
        type: number
      paymentsApplied:
        type: string
      balanceDue:
        type: number
      laborCost:
        type: number
      equipmentCost:
        type: number
      materialCost:
        type: number           
      stageId:
        type: string
      taxRate:
        type: number
      billingAddress1:
        type: string
      billingAddress2:
        type: string
      billingCity:
        type: string
      billingState:
        type: string
      billingZip:
        type: number
      shippingAddress1:
        type: string
      shippingAddress2:
        type: string
      shippingCity:
        type: string
      shippingState:
        type: string
      shippingZip:
        type: number           
      otherCost:
        type: number     
      createdBy:
        type: string
  updateInvoice:
    type: object
    properties:
      invoiceId:
        type: string  
      contactId:
        type: string
      projectId:
        type: string
      individualId:
        type: string
      opportunityId:
        type: string
      orderId:
        type: string
      estimateId:
        type: string
      poNumber:
        type: string
      title:
        type: string
      createdDate:
        type: string
      dueDate:
        type: string
      invoiceNumber:
        type: string
      invoiceItem:
        type: string
      terms:
        type: number
      class:
        type: number
      salesRep:
        type: string
      customerMessage:
        type: number
      memo:
        type: string
      total:
        type: number
      paymentsApplied:
        type: string
      balanceDue:
        type: number
      laborCost:
        type: number
      equipmentCost:
        type: number
      materialCost:
        type: number           
      stageId:
        type: string
      taxRate:
        type: number
      billingAddress1:
        type: string
      billingAddress2:
        type: string
      billingCity:
        type: string
      billingState:
        type: string
      billingZip:
        type: number
      shippingAddress1:
        type: string
      shippingAddress2:
        type: string
      shippingCity:
        type: string
      shippingState:
        type: string
      shippingZip:
        type: number           
      otherCost:
        type: number     
      modifiedBy:
        type: string
  deleteInvoice:
    type: object
    properties:
      invoiceId:
        type: string
  getInvoiceDetails:
    type: object
    properties:
      invoiceId:
        type: string
...